+++ router/core/router.go
@@ -1660,6 +1660,7 @@ func WithApolloCompatibilityFlagsConfig(cfg config.ApolloCompatibilityFlags) Opt
 	return func(r *Router) {
 		if cfg.EnableAll {
 			cfg.ValueCompletion.Enabled = true
+			cfg.TruncateFloats.Enabled = true
 		}
 		r.apolloCompatibilityFlags = cfg
 	}